requirements.txt
discord.py==2.4.0 
psycopg2-binary==2.9.9 
python-dotenv==1.0.1 
aiohttp==3.9.5
psutil==5.9.8
asyncpg==0.29.0
google-cloud-secret-manager==2.20.0
reportlab==4.0.4